+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Improvements
+- Add the "desc" field (check's description) to API responses
+
+
 ## 1.5.0 - 2019-02-04
 
 ### Improvements